refactor: frontend static resource with dry run

+++ src/veil/server/frontend/frontend.py
@@ -10,6 +10,10 @@ def frontend_static_resource(frontend_root_path):
     frontend_root_path = as_path(frontend_root_path)
     if not frontend_root_path or not (frontend_root_path / 'package.json').exists():
         return
+    dry_run_result = get_dry_run_result()
+    if dry_run_result is not None:
+        dry_run_result['frontend_static_resource'] = 'INSTALL'
+        return
     shell_execute('sudo npm install yarn -g --registry=https://registry.npm.taobao.org', cwd=frontend_root_path)
     shell_execute('yarn install --registry=https://registry.npm.taobao.org', cwd=frontend_root_path)
     shell_execute('yarn run build', cwd=frontend_root_path)
